question_generation/requirements.txt
PyMuPDF==1.24.2
beautifulsoup4==4.12.3
langchain==0.1.16
langchain_community==0.0.34
langchain_core==0.1.46
langchain_openai==0.1.4
langchain_text_splitters==0.0.1
loguru==0.7.2
openai==1.24.0
python-dotenv==1.0.1
PyYAML==6.0.1
Requests==2.31.0
streamlit==1.32.0
python-multipart==0.0.9
fake-useragent==1.5.1
ffmpeg-python==0.2
Django==4.2.11
pymongo==4.7.2
mongoengine==0.28.2
docx2txt==0.8
torch==2.3.0
transformers==4.41.0
celery==5.4.0
redis==3.5.3
flower==2.0.1
slack-sdk==3.30.0